updateSelectOptions 함수 분리
