+++ src/basic/components/index.js
@@ -1,0 +1,3 @@
+export * from "./bonusPoints.js";
+export * from "./updateSelectOptions.js";
+export * from "./updateStock.js";
